[Behat][Checkout] Raw moving payment steps
