+++ src/Sylius/Behat/Context/Ui/Shop/PaymentCheckoutContext.php
@@ -1,0 +1,129 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Behat\Context\Ui\Shop;
+
+use Behat\Behat\Context\Context;
+use Sylius\Behat\Page\Shop\Checkout\SelectPaymentPageInterface;
+use Sylius\Behat\Service\SharedStorageInterface;
+use Webmozart\Assert\Assert;
+
+/**
+ * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
+ */
+final class PaymentCheckoutContext implements Context
+{
+    /**
+     * @var SharedStorageInterface
+     */
+    private $sharedStorage;
+
+    /**
+     * @var SelectPaymentPageInterface
+     */
+    private $selectPaymentPage;
+
+    /**
+     * @param SharedStorageInterface $sharedStorage
+     * @param SelectPaymentPageInterface $selectPaymentPage
+     */
+    public function __construct(SharedStorageInterface $sharedStorage, SelectPaymentPageInterface $selectPaymentPage)
+    {
+        $this->sharedStorage = $sharedStorage;
+        $this->selectPaymentPage = $selectPaymentPage;
+    }
+
+    /**
+     * @When I decide to change order shipping method
+     */
+    public function iDecideToChangeMyShippingMethod()
+    {
+        $this->selectPaymentPage->changeShippingMethod();
+    }
+
+    /**
+     * @When /^I choose "([^"]*)" payment method$/
+     */
+    public function iChoosePaymentMethod($paymentMethodName)
+    {
+        $this->selectPaymentPage->selectPaymentMethod($paymentMethodName ?: 'Offline');
+        $this->selectPaymentPage->nextStep();
+    }
+
+    /**
+     * @Given I am at the checkout payment step
+     * @When I go back to payment step of the checkout
+     */
+    public function iAmAtTheCheckoutPaymentStep()
+    {
+        $this->selectPaymentPage->open();
+    }
+
+    /**
+     * @When I complete the payment step
+     */
+    public function iCompleteThePaymentStep()
+    {
+        $this->selectPaymentPage->nextStep();
+    }
+
+    /**
+     * @When I select :paymentMethodName payment method
+     */
+    public function iSelectPaymentMethod($paymentMethodName)
+    {
+        $this->selectPaymentPage->selectPaymentMethod($paymentMethodName);
+    }
+
+    /**
+     * @Then I should be on the checkout payment step
+     */
+    public function iShouldBeOnTheCheckoutPaymentStep()
+    {
+        Assert::true(
+            $this->selectPaymentPage->isOpen(),
+            'Checkout payment page should be opened, but it is not.'
+        );
+    }
+
+    /**
+     * @Then I should not be able to select :paymentMethodName payment method
+     */
+    public function iShouldNotBeAbleToSelectPaymentMethod($paymentMethodName)
+    {
+        Assert::false(
+            $this->selectPaymentPage->hasPaymentMethod($paymentMethodName),
+            sprintf('Payment method "%s" should not be available, but it does.', $paymentMethodName)
+        );
+    }
+
+    /**
+     * @Then I should be able to select :paymentMethodName payment method
+     */
+    public function iShouldBeAbleToSelectPaymentMethod($paymentMethodName)
+    {
+        Assert::true(
+            $this->selectPaymentPage->hasPaymentMethod($paymentMethodName),
+            sprintf('Payment method "%s" should be available, but it does not.', $paymentMethodName)
+        );
+    }
+
+    /**
+     * @Then I should be redirected to the payment step
+     */
+    public function iShouldBeRedirectedToThePaymentStep()
+    {
+        Assert::true(
+            $this->selectPaymentPage->isOpen(),
+            'Checkout payment step should be opened, but it is not.'
+        );
+    }
+}
